updated script to be linux-agnostic

+++ setup.sh
@@ -1,38 +1,63 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
-## setup ubuntu/debian in wsl
+LX_ENV=any
 
-git init -b wsl &&
-    git remote add origin https://github.com/$USER/dotfiles.git &&
-    git pull origin wsl
+# check if WSL
+if [ -n "$WSL_DISTRO_NAME" ]; then
+    LX_ENV=wsl
+fi
 
-# install nala
-sudo apt update &&
-    sudo apt install -y nala
+# check if NixOS
+if [ -e /etc/nixos ]; then
+    LX_ENV=nixos
+fi
 
-# update distro
-sudo nala update &&
-    sudo nala upgrade -y
+echo "installing setup for $LX_ENV"
 
-# nix + home manager
-curl -fsSL https://nixos.org/nix/install | bash /dev/stdin --no-daemon &&
-    mkdir -p $HOME/.config/nix &&
-    echo "experimental-features = nix-command flakes" >> $HOME/.config/nix/nix.conf &&
-    mv "$HOME"/.bashrc "$HOME"/.bashrc_default &&
-    mv "$HOME"/.profile "$HOME"/.profile_default &&
-    . "$HOME"/.nix-profile/etc/profile.d/nix.sh &&
-    curl "https://raw.githubusercontent.com/cethien/setup/refs/heads/wsl/resources/config.template.json" >> $HOME/.config/home-manager/config.json &&
+# for non nixOS
+if [ "$LX_ENV" = "any" ]; then
+    # install nala for for distros with apt and update
+    if command -v apt >/dev/null 2>&1; then
+        echo "installing nala"
+        sudo apt update &&
+            sudo apt install -y nala curl &&
+            sudo nala update &&
+            sudo nala upgrade -y
+    fi
+
+    # install nix
+    echo "installing nix"
+    curl -fsSL https://nixos.org/nix/install | bash /dev/stdin --no-daemon &&
+        mkdir -p $HOME/.config/nix &&
+        curl https://raw.githubusercontent.com/cethien/setup/lx/resources/nix.conf >>$HOME/.config/nix/nix.conf &&
+        mv "$HOME"/.bashrc "$HOME"/.bashrc_default &&
+        mv "$HOME"/.profile "$HOME"/.profile_default &&
+        . "$HOME"/.nix-profile/etc/profile.d/nix.sh
+fi
+
+$script = '
+git init -b lx &&
+    git remote add origin https://github.com/cethien/dotfiles.git &&
+    git reset --hard origin/lx &&
+    git pull --set-upstream origin lx &&
+    curl https://raw.githubusercontent.com/cethien/setup/lx/resources/config.template.json >>$HOME/.config/home-manager/config.json &&
     nano $HOME/.config/home-manager/config.json &&
-    nix build .config/home-manager#homeConfigurations.$USER.activationPackage &&
-    result/activate &&
-    home-manager switch &&
-    rm -rf result
+    home-manager switch
+'
 
-# add windows ssh-passthru
-curl -fsSL https://gist.githubusercontent.com/Jaykul/19e9f18b8a68f6ab854e338f9b38ca7b/raw/ssh-agent-pipe.sh | sudo tee /usr/local/bin/ssh-agent-pipe >/dev/null &&
-sudo chmod +x /usr/local/bin/ssh-agent-pipe
+nix-shell -p git home-manager nano curl --run "$script"
 
-# remove sudo pw prompt
-echo "$USER ALL=(ALL) NOPASSWD:ALL" | sudo tee /etc/sudoers.d/"$USER" >/dev/null
+# for wsl,
+if [ "$LX_ENV" = "wsl" ]; then
+    # add ssh-passthrough script
 
-sudo reboot
+    echo "adding ssh-passthrough script"
+    curl -fsSL https://raw.githubusercontent.com/cethien/setup/lx/resources/ssh-agent-pipe.sh | sudo tee /usr/local/bin/ssh-agent-pipe >/dev/null &&
+        sudo chmod +x /usr/local/bin/ssh-agent-pipe
+
+    # remove sudo password prompt
+    echo "$USER ALL=(ALL) NOPASSWD:ALL" | sudo tee /etc/sudoers.d/"$USER" >/dev/null
+
+    # reboot
+    sudo reboot
+fi
